Reorganise coda functions into separate files for each sync direction
Also fixes the name of the coda module to reflect that it contains files for syncing in both directions
