+++ src/pipeline_configuration_spec.py
@@ -5,7 +5,7 @@ from core_data_modules.data_models import CodeScheme
 
 from src.common.configuration import RapidProClientConfiguration, CodaClientConfiguration, UUIDTableClientConfiguration, \
     EngagementDatabaseClientConfiguration
-from src.engagement_db_to_coda.configuration import CodaSyncConfiguration, CodaDatasetConfiguration, \
+from src.engagement_db_coda_sync.configuration import CodaSyncConfiguration, CodaDatasetConfiguration, \
     CodeSchemeConfiguration
 from src.engagement_db_to_rapid_pro.configuration import EngagementDBToRapidProConfiguration, DatasetConfiguration, \
     WriteModes, ContactField
